requirements-scraper.txt
# Twitter Scraper Dependencies v1.0
playwright>=1.40.0
python-dotenv>=1.0.0
asyncio-throttle>=1.0.2

# Database
# sqlite3 is usually included with Python

# HTML Parsing and Data Extraction
beautifulsoup4>=4.12.0
lxml>=4.9.0

# GUI Dependencies 
# tkinter is usually included with Python
# If tkinter is not available, install python-tk on Linux systems

# Optional: For better logging and data handling
pandas>=2.0.0
